requirements.txt
torch
torchaudio
mpi4py
tqdm
transformers
llama-cpp-python
spacy
pandas
numpy